+++ wcomponents-core/src/main/java/com/github/bordertech/wcomponents/WTable.java
@@ -44,7 +44,7 @@ import org.apache.commons.logging.LogFactory;
  * @author Jonathan Austin
  * @since 1.0.0
  */
-public class WTable extends WBeanComponent implements Container, AjaxTarget, SubordinateTarget,
+public class WTable extends WBeanComponent implements Container, AjaxInternalTrigger, AjaxTarget, SubordinateTarget,
 		Marginable, NamingContextable {
 
 	/**
@@ -909,6 +909,7 @@ public class WTable extends WBeanComponent implements Container, AjaxTarget, Sub
 
 	/**
 	 * Set the table as having row headers.
+	 *
 	 * @param rowHeaders indicates that the first data column in the table should be considered a row header.
 	 */
 	public void setRowHeaders(final boolean rowHeaders) {
@@ -1497,23 +1498,6 @@ public class WTable extends WBeanComponent implements Container, AjaxTarget, Sub
 			}
 
 			return parsed.toArray(new String[parsed.size()]);
-		}
-	}
-
-	/**
-	 * Override preparePaint to register an AJAX operation if necessary.
-	 *
-	 * @param request the request being responded to.
-	 */
-	@Override
-	protected void preparePaintComponent(final Request request) {
-		super.preparePaintComponent(request);
-
-		if (getRowsPerPageOptions() != null || PaginationMode.DYNAMIC.equals(getPaginationMode()) || SortMode.DYNAMIC.
-				equals(getSortMode())
-				|| ExpandMode.DYNAMIC.equals(getExpandMode()) || ExpandMode.LAZY.equals(
-				getExpandMode())) {
-			AjaxHelper.registerComponentTargetItself(getId(), request);
 		}
 	}
 
@@ -2233,7 +2217,7 @@ public class WTable extends WBeanComponent implements Container, AjaxTarget, Sub
 	 * </p>
 	 * <p>
 	 * Note that Data may be stored locally or sourced remotely, depending on the particular TableModel implementation.
-	 * <p>
+	 * </p>
 	 * <p>
 	 * The row indexes used in the interface are a list of row indexes. Each item in the list is the index of the row
 	 * for that level. The size of the list passed in matches the depth of the row.