+++ wcomponents-theme/src/main/js/wc/dom/ariaAnalog.js
@@ -371,11 +371,12 @@ define(["wc/has",
 
 			if (event.canCapture) {
 				event.add(element, event.TYPE.focus, eventWrapper.bind(this), null, null, true);
+				event.add(element, event.TYPE.click, eventWrapper.bind(this), null, null, true);
 			}
 			else {
 				event.add(element, event.TYPE.focusin, eventWrapper.bind(this));
+				event.add(element, event.TYPE.click, eventWrapper.bind(this));
 			}
-			event.add(element, event.TYPE.click, eventWrapper.bind(this));
 			event.add(element, event.TYPE.keydown, eventWrapper.bind(this));
 			shed.subscribe(shed.actions.SELECT, this.shedObserver.bind(this));
 			shed.subscribe(shed.actions.DESELECT, this.shedObserver.bind(this));
@@ -632,7 +633,6 @@ define(["wc/has",
 		 */
 		AriaAnalog.prototype.activate = function(element, SHIFT, CTRL) {
 			var container,
-				lastActivated,
 				selectMode,
 				isMultiSelect,
 				setLastActivated = true;